requirements.txt
openai
questionary
python-dotenv
Pillow
requests
